refactor toolbar code
This groups toolbar buttons by function, and puts buttons for each
group in an alist. In turn, button groups are collected into a
singleton alist called toolbar-button-groups, which is visible in
the global namespace. Hence, the individual button names do not
pollute the global namespace anymore.

The toolbar-button procedure now also generates an info string, which
is shown in the status bar on mouse <Enter> events. It temporarily
replaces the active command info, which is restored on mouse
<Leave> (if any).