fix keypress evaluation
Tk's default behaviour is to first attempt and match bindings specific
to a widget. This causes global key bindings like <Control-x> to be
matched to a less specific widget-local binding such as <Keypress-x>
first, which is obviously not what we want.

This commit fixes the problem by applying 2 strategies.

1) Provide a procedure to reverse binding evalutation order for a
widget with `bindtags`.

2) Prevent event propagation for global keypress events by appending a
break command with `bind . <event-spec> +break`.