quick hack for fixing numeric entry
The <<BlockEntry>> event assumes only keys defined in key group
note-edit to be valid. This quickfix ensures that keys 0-9 and a-f are
always recognized as valid entry keys. However, this is a problem
since it will not work with azerty, nor will it work if the default
number base is > 16.