MDAL renaming, part 1
We have too many "configs". To make things clearer, MDAL
Configurations are re-branded as MDAL Engine Definitions.

This commit updates naming in the MMOD and MDEF standard
specifications, and improves some structure/wording in these documents.

The following commits will change the file endings to .mdef and .mmod,
and rename things internally to consistently use the MMOD and MDEF
abbreviations. This includes renaming MDMOD to MMOD.