Schemta rewrite
This gets rid of global state in the assembler for the most
part. Bindings required to evaluate instruction sets and s-expression
directives in asm code are now bound dynamically.

The `assemble` procedure is now a high-level abstraction over
`make-assembly`. The latter parses a source on creation, returning a
closure that manages the assembly process' state. The closure can be
called repeatedly to do 1 or more passes until all nodes in the ast
are resolved.

This is important for the MDAL compiler: It's the basis for an
upcoming commit that will change the way asm output nodes
work. Currently asm nodes must be evaluated on initialization, which
means asm code cannot utilize symbols that are created dynamically
during module compilation, for example loop points. The new Schemta
design makes it possible to only evaluate partially on init and defer
the final assembly pass(es) until the module is compiled.