fix "??" field conditions, implement dynamic oblock fields
Previously, a field node id prefixed with `??` evaluated to a
procedure that returned `#t` when the relevant field node instance
evaluated to a non-null value. This commit changes the behavior so
that `#t` is only returned if the for field node instances that are
set (ie, the old `is-set?` condition).

Dynamic oblock fields carry an additional condition specifier. Field
values are only emitted when the condition holds true. This is useful
for row data compression schemes commonly found in newer 1-bit
engines.