fix hanging button hover highlights
Calling `(tk/update 'idletasks)` on Enter/Leave events and not using a
lock seems to do the trick.

However, there is a bit of a risk associated with this fix. In order
to make this work, I've disabled tk/safe-dialogue for opening files,
as that would fail to bring the file selector menu to the foreground
after a previous dialog ran (eg. the "save changes?" dialog). However,
I originally implement tk/safe-dialogue for the very same reason. So
we'll need to keep an eye on how well this works now.