fix file selector dialog styling, reindent
This affects *nix only, on Windows native dialog widgets are used.

The file name field is a ttk::entry, so it can be styled by providing
defaults for TEntry.

The list of files is a ::tk::IconList, which pulls color information
for selected entries (and only for those) from the `focus` dynamic
state of TEntry.

Unfortunately the background of the list of files is a canvas widget,
which is not affected by the Tk theme settings.