implement dwim-module-edit
Do-what-I-mean module edit attempts to adjust values that are to be
pasted/inserted into a block, if those values do not meet the
constraints of the target field. If the target field uses a key/ukey
command, and the source values are integers, then those integers are
scaled to the target range and then converted to the closest matching
key. If the target field uses an int/uint command, and the source
values exceed the target range, then the source values are scaled
accordingly. If the source values are keys, they are converted to
their respective values and then scaled. In any other case, source
values are returned unchanged.

The main purpose of this is to allow conversion between notes and
numeric fields.

Dwim-module-edit can be enabled or disabled with the
`dwim-module-edit` setting. It is enabled by default.