another attempt at fixing crashes on fast scrolling
This commit does two things:

1. Raise the (fairly conservative) recursion limit of the Tcl
interpreter from 1000 to 20000. Not a particular clean solution, but
should be safe enough. The comment below this
[accepted answer](https://stackoverflow.com/questions/50340318/continuing-a-loop-for-a-proc-within-a-proc)
on stackoverflow suggests that limits of >1000000 are fine, and 20000
should be well within limits even for older machines which impose
tighter restrictions on C stack size.

2. Force an update after modifying textgrid tags. Textgrid tag updates
may generate a large number of calls which end up getting nested due
to how pstk works. Forcing an update after a run of tag updates should
impose a soft cap on the maximum number of nested calls that can occur.