rework dialogs, add `<ui-wrapper>` class
The original closure-based dialog implementation was written before
the change to an object oriented GUI system, and did not integrate
well with the latter. This commit reimplements dialogs as a coops
class. With this change, it becomes possible to utilize Bintracker's
high level GUI API in dialogs.

Also provided is a `<ui-wrapper>` class, which makes it possible to
use generic Tk widgets as Bintracker UI class instances.

Code using the new <ui-dialog> class will be a bit more verbose, but
also more powerful than code using the old implementation.

The new dialog and ui-wrapper code is quite experimental at this
point, and will likely require more tweaks in the future.