rework pseudo-random number generators
- All prng procedures now accept a MAXINT argument, instead of a BITS
argument. This obviously makes the generators more flexible; as a
consequence it is no longer necessary to scale results for a desired
range where the maximum is not a power of 2. The unzufall plugin is
adjusted accordingly.
- The Blum-Blum-Shub generator now properly generates a bit sequence,
rather than exposing internal state as result.
- A new `stat` module transforms uniformly distributed results from
PRNGs into non-uniform ones. Currently it only supports standard
normal (Gaussian) distribution.