add AMP16k engine (preliminary support)
This adds Hikaru's multi-channel engine for the ZX Spectrum 16K. A
number of problems currently remain:

- libmdal does not yet support per pattern single fields, so tempo is
currently fixed to 6 ticks/per row
- AMP16k requires the duty field to be forced on the first note of a
channel. However, it will ignore the duty setting if done on a
rest. This makes it difficult to support in libmal, so we currently
rely on the user setting the duty manually.
- The ZX core in MAME frequently fails to set PC on this engine,
requiring a user reset. This is very bad since it isn't obvious that
this is happening.
- MAME fails the contention test, so we have to run from uncontended
memory.