use modifiers for interpolation, split off bt-maths module
On interpolating key/ukey fields that have modifiers enabled, check if
the associated modifier column is selected. If it is, then a) base the
interpolation on the modified values, and b) write resulting offsets
from the key/ukey column into the modifier column.

This commit splits off some helper code from the `ui-interpolate`
method into a new "bt-maths" module.