ensure LUA bridge compatibility with MAME >= 0.227
In MAME version 0.227, the LUA engine was streamlined, introducing a
number of backwards-incompatible changes. Namely, the
`manager.machine` property replaces the old `manager:machine()`
method.

This commit introduces a work-around to ensure our LUA bridge remains
compatible with both pre- and post-0.227 MAME. It works by reading the
MAME version number, and assigning either to `manager.machine` or
`manager:machine()` to an abstract `machine_manager` local, depending
on the result of the version check.

This resolves #8.