make specifying an origin address optional for module compilation
The default use case is to compile modules at their default address,
which can be determined from the module's engine definition. So it is
more convenient to make the origin argument an optional keyword, so
the user can eg. call `mod->bin` without having to figure out a valid
origin address.

This affects the compilation procedures `mod-export-bin`,
`mod-export-asm`, `mod->asm`, and `mod->bin`.