use abstract symbolic emulator names
Previously, system->emulator lookup used the name of the target
emulator executable as lookup key. However, the idea behind having
this lookup in the first place is to provide a single place for
specifying how the executable for a certain emulator (usually MAME) is
called. To do this properly, we now use a symbolic name for the
emulator argument in `config/systems.scm`, which is used as a key to
look up the actual executable name and additional command line
arguments in `config/emulators.scm`. So to use an emulator executable
with a different name, only the entry in the latter file needs to be
changed now.