improve MAME interaction
This commit contains various improvements to Bintracker's interaction
with MAME. `mame_bridge.lua` is reworked to be more stable, and the
bt-emulation API is simplified.

Machine specific parameters (names of registers, loader type, default
run addresses, etc.) are now handled on the MAME side. This means it
is no longer necessary to specify these details in
`config/systems.scm`, or pass them as arguments when creating an
emulator instance directly through `make-emulator`.

Furthermore, a long-standing stability issue that was causing audio
glitches on row play is resolved. Apparently, loading a new binary
into memory and then running it will corrupt memory on ZX Spectrum, if
the stack pointer points to an address within the target region of the
load. To solve this, mame_bridge now sets the stack pointer to a safe
address on loading a binary. There are still similar issues on other
target platforms. Those will be addressed in a future commit.