fix unreliable ChannelF playback
When ChannelF emulator core has a jump instruction cached, it will
execute it, ignoring writes to PC0/CURPC/GENPC. The work-around is to
perform a soft reset after loading a new binary. This is not ideal,
since calling soft_reset() is somewhat slow, but the delay seems
tolerable, so... good enough for now.