check if emulator exists before attempting to run it
This fixes a crash that would occur when trying to run an emulator
that does not exist on the user's system.

Attempting to run a non-existing emulator with `make-emulator` will
raise an exception. `<module-view>`'s constructor catches this, and
disables playback functions accordingly, but still allows the user to
edit the module (albeit silently).