implement basic support for fixed-size blocks and reference commands
This is the first step towards implementing support for effects
tables, wavetables, samples, etc. For now, only backward references
are handled correctly (ie. a node must be evaluated before it can be
referenced), forward references will crash the compiler.

Implementing reference commands raised an interesting problem. Order
lists (song sequences) are implemented with reference
commands. I generally consider this to be the correct approach,
because entries in an order list are in fact references (to pattern
blocks). However, we do not want these references to be resolved
during the compilation of an order block (ie. we don't care about the
underlying memory addresses while we evaluate an order list), because
order nodes have an additional post-evaluation step, during which they
are resolved in varying ways depending on the desired output data layout.

For now, I implemented a quick and dirty solution. When an order block
is evaluated, a flag is passed to `eval-block-field`, which causes
reference fields to be evaluated as unsigned integers values
instead. However, aside from being an ugly design, this is problematic
because it prevents the use of actual reference commands in order
blocks (a possible use case would be patterns that reference some
effects table on a per-pattern basis).

I'm thinking about two possible options at the moment. The easier
solution would be to split reference commands into "numeric" and
"symbolic" references. The GUI would treat both of these the same way,
but the compiler would only evaluate symbolic references, and not
numeric ones. So order references would then become numeric
references. However, I'm not too psyched about introducing an extra
command type in MDEF.

A more elegant, but much more work-intensive solution would be to
replace the entire compiler generator with a more intelligent design,
that separates evaluation and symbolic resolution more cleanly. I'm
leaning towards this solution at the moment, however I would prefer
holding off on actually doing this, and gather some more insight into
what's needed first. So, implementation might have to wait until after
the beta release. Generally, this is not a problem, as the compiler
generator is well isolated from the rest of the code. So for now,
quick and dirty it is.