prepare for implementation of unordered MMOD groups
This is mostly a bag of different tweaks und small
refactorings. However, this is actually sufficient to implement
unordered groups on a basic level.

While implementing the initial design for unordered groups, I realized
that actually the most elegant solution is to not implement anything.

The idea is to always create order lists for group nodes, regardless
if they are used or not. This way, a lot of special handling code can
be avoided. There are only two components that need to distinguish
between ordered and unordered groups, the compiler and the GUI. These
changes will be implemented in upcoming commits.

The compiler needs to take order into account when re-arranging block
instances by merging and splitting. For ordered groups, blocks should
be merged according to order. For unordered groups, blocks should be
merged according to the list of unique positions in the order.

For the GUI, `<ui-order-view>` needs to hide certain fields from view
for unordered groups, and disable editing of some other fields. For
unordered groups with only one block member and a fixed block
size, only inserting/deleting/appending/removing of
nodes should be available. I think this can wait until simplified
orders are implemented, because those would add most of the code
required for this.

The downside is that this makes the data format counter-intuitive in
one specific scenario. Consider a group that has one fixed-length
block member (eg. a wavetable). From the user's point of view it makes
no sense having to write an order node for this. It's also easy to
automatically generate an order in this case, so that's a thing that
should be implemented at some point.

However, having an order for "unordered" nodes does actually make
sense even for groups with a fixed block size. This becomes apparent
when there is more than one block member. Since orders are virtual in
MDAL, blocks must be merged during compilation. That is, all of the
blocks' fields are combined into one super-block. With multiple
blocks, this must happen in sequence according to the user's
intentions. Therefore, an input order is needed even though the output
is unordered.