hide length field in order-view when block-length is fixed
This is part of the ongoing effort to implement the GUI for
samples/tables etc.

Using a general-purpose solution here, which makes it possible to hide
a field column in `<ui-basic-block-view>`. Besides hiding the "ROWS"
field in order views, this will also be useful for implementing
toggles to show/hide selected columns later on.

However, this commit changes a number of commonly used low-level
methods, which will require more testing before it becomes fire-proof.