re-implement compiler/compiler generator
This drastically simplifies the compiler at MDAL level, offloading
most of the heavy lifting to the Schemta assembler.

The MDAL compiler no longer tracks the output origin address. Instead,
it produces a fully position-independent abstract syntax tree for
Schemta ahead of time (which allows for more aggressive caching to be
implemented later on). At runtime, the compiler produces just a list
of symbols, which Schemta will use to resolve the AST. This makes it
possible to evaluate output nodes out of order, avoiding multiple
passes by the MDAL compiler at the cost of (generally 1-2) more
assembler passes.

Deferring address resolution in this manner is a prerequisite for
implementing the optimizer, since we can now replace and remove output
node content without worrying about breaking references. Deferred
address resolution is also needed for handling MDAL label commands.

The new implementation should be faster in most cases, but will
currently be slower than before when the output contains many
cross-references, such as in the case of the wiz4v engine. This issue
will be addressed in a further commit.