prepare for windows builds
Builds and boots up on Wine, at least. I expect there are still plenty
of problems to be solved before Bintracker actually works on Windows,
though.

Also, thanks to Kristian Lein-Mathisen for helping to fix an issue
with stb-image, caused by not reading image files in binary mode on
Windows.