more changes for windows builds, use msys2 instead of plain mingw
There was a problem with plain mingw buils (both cross-compiled and
compiled natively on Windows), where builds would either have
insufficient stack limits (manifesting as exception 0xC0000028
STATUS_BAD_STACK on actual Windows) or cause stack overflows.

The only setup I was able to make work so far involves compiling on
actual Windows, using msys2 and chicken installed through
Chocolatey. That process still needs quite a bit of massaging (as
documented in docs/setup.md). The builds from this setup don't crash
at least, but there is an unresolved issue with the emulator thread
blocking execution of the main program.

There was also an issue with populating the database. Turns out that
the `directory` procedure from `chicken.file` is broken in
Windows. This commit provides a work-around by using `capture` from
the `shell` egg to run a directory listing on the command line.