fix bitx engine playback
Starting with
https://github.com/mamedev/mame/commit/0f17b85baf6e5f396c14f7188904488355bdf607,
16K RAM is no longer enabled by default for TRS MC-10, which breaks
playback for the bitx engine.

This can be remedied by passing an extra command line option on
startup. However, this option is not available on MAME versions before
the above commit, which means this fix will not work on MAME < 0.233.

Since a) this affects only one machine and b) Bintracker will drop
support for all MAME versions < 0.241 soon, I believe it's not worth
it to provide a messy work-around just for this.