add PTC Sol-20 support
Unfortunately it's a bit wonky since users need to manually dismiss
MAME's warning about the machine not working properly. I wish there
was an option in MAME to skip those warnings, but unfortunately
MAMEdev has ruled out ever adding that. Especially since in this case,
it complains about a ROM that doesn't exist anywhere and isn't
required to run the machine.