explicitly mark MDAL groups as `playable`
For the moment, the main purpose of this is to provide a reliable
mechanism for detecting whether Pattern Play should do something or
not, depending on which MDAL group is currently in view. Currently
only works for regular patterns. Could later be extended to call a
group-specific playback function (for example for previewing samples).