research support for emulators with gdb remote debugging
Works on a basic level, but I currently don't have an emulator at hand
that supports enough of the gdb protocol to be useful for Bintracker.
Tried with xroar but that one turns out to be to simple - neither
reset, start, stop, nor quit actions are supported.

I'm therefore putting this on hold while trying to clean up the
current source code mess.