mame bridge: enable use of write taps for halting machines
Write taps are a feature in MAME Lua scripting that allows (among
other things) triggering a Lua callback when the emulated code writes
to a given memory location. We can use this to halt the emulated
machines on single row/one-shot playback.

Write taps are not without flaws. They may fire during
startup/reset (which can mostly be worked around), and, more
critically, may not always fire when they're supposed to. Halting
machines is still somewhat more reliable for some systems (CoCo3,
Dragon) than halting via the "loop and spam the virtual instruction
register with NOPs" approach we're currently using.

As write taps are a recent addition to MAME, let's use the opportunity
to also drop backwards compatibility with MAME <0.227.