reduce usage of eval in asm instruction resolution
Bind instruction operands dynamically, instead of EVALuating them.
This speeds up instruction resolution by ~20%.

The new solution limits the number of possible operands to 3, instead
of supporting an infinite number of operands as before. Since we
aren't messing with x86 assembly, this should be fine, though the
solution can easily be extended to accomodate for more operands if needed.